printf expects a unsigned long no matter the archecture.

+++ basestation/src/anki/cozmo/basestation/actions/driveToActions.cpp
@@ -576,7 +576,7 @@ namespace Anki {
       
       PRINT_NAMED_INFO("DriveToPoseAction.SetGoal",
                        "Setting %lu possible goal options.",
-                       _goalPoses.size());
+                       (unsigned long)_goalPoses.size());
       
       _isGoalSet = true;
